Added detailed messages to settings cloning
Added ability to eject unpackager trackers
Added ability to clear recipe holders and markers in crafting grid
Added settings cloning to fluid package filler
Changed saving settings to cloner to use shift-right click

+++ src/main/java/thelm/packagedauto/integration/emi/EncoderDragDropHandler.java
@@ -86,7 +86,7 @@ public class EncoderDragDropHandler implements EmiDragDropHandler<EncoderScreen>
 		public void accept(EmiIngredient emiIngredient) {
 			ItemStack stack = wrapStack(emiIngredient);
 			if(!stack.isEmpty()) {
-				PacketHandler.INSTANCE.sendToServer(new SetItemStackPacket((short)slot.index, stack));
+				PacketHandler.INSTANCE.sendToServer(new SetItemStackPacket(slot.index, stack));
 			}
 		}
 	}
